BlockTable 구현 및 기존 테스트에 적용
