+++ src/main/kotlin/minesweepertest/BlockTable.kt
@@ -1,0 +1,5 @@
+package minesweepertest
+
+class BlockTable(
+    val record: Map<Pair<Int, Int>, Block>
+)
