+++ src/main/kotlin/minesweepertest/GameMap.kt
@@ -1,5 +1,5 @@
 package minesweepertest
 
 class GameMap(
-    val blocks: Map<Pair<Int, Int>, Block>
+    val blockTable: BlockTable
 )
